added channel list on sidebar
