+++ src/js-component/components.css
@@ -1,6 +1,6 @@
 
 
-/* message bar */
+/* message layout */
 .message_bar {
     width: 95%;
     border: 1px solid white;
@@ -29,6 +29,14 @@
     color: white;
 }
 
+.message_interface {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 70%;
+    align-self: flex-end;
+    margin-bottom: 30px;
+}
 
 /* sidebar & memberlist */
 .sidebar {
@@ -47,12 +55,15 @@
     display: flex;
     justify-content: space-between;
 }
-.sidebar button {
+.plus_buttons, .cdm_buttons{
     background-color: rgb(29,97,194);
     border: none;
     color: white;
     font-size: 30px;
     cursor: pointer;
+}
+.cdm_buttons {
+    font-size: 16px;
 }
 
 /* message log */
