+++ src/pages/SlackInterface.js
@@ -1,10 +1,9 @@
 import MemberList from "../js-component/MemberList";
-import MessageBar from "../js-component/MessageBar";
-import MessageLog from "../js-component/MessageLog";
 import Sidebar from "../js-component/Sidebar";
 import CreateChannel from "../js-component/CreateChannel";
 import './interface.css'
 import { useState } from "react";
+import MessageLayout from "../js-component/MessageLayout";
 
 const SlackInterface = () => {
     const [popup, setPopup] = useState(false)
@@ -16,10 +15,7 @@ const SlackInterface = () => {
         <>
             <div className="slack_interface">
                 <Sidebar showPopup={showPopup} />
-                <div className="message_interface">
-                    <MessageLog />
-                    <MessageBar />
-                </div>
+                <MessageLayout />
                 <MemberList />
             </div>
             <CreateChannel popup={popup} showPopup={showPopup}/>
