+++ src/pages/interface.css
@@ -13,12 +13,3 @@ body {
     flex-direction: row;
     height: 100vh;
 }
-
-.message_interface {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 70%;
-    align-self: flex-end;
-    margin-bottom: 30px;
-}
